+++ src/sponge/mod.rs
@@ -4,7 +4,8 @@
 //
 // Copyright (c) DUSK NETWORK. All rights reserved.
 
-mod sponge;
+pub mod sponge;
+pub mod truncated;
 
 pub use sponge::sponge_gadget as gadget;
 pub use sponge::sponge_hash as hash;